fix imports and create argparser function and main file to use main function to run the tool
