+++ yaseer_tool/functions/consts.py
@@ -8,7 +8,26 @@ DATE = datetime.now().time()
 COLOR = colorama.Fore
 
 
-def ssh_login_data(user, password, hostname, port) -> str:
+def thePath(path):
+    return f"{COLOR.WHITE}[{COLOR.BLUE}{DATE}{COLOR.WHITE}]{COLOR.GREEN} {path}"
+
+
+def tcp(port, name, product, version, extrainfo):
+    return f"\r{COLOR.WHITE}[{COLOR.BLUE}{DATE}{COLOR.WHITE}] {COLOR.WHITE}[{COLOR.GREEN}OPEN{COLOR.WHITE}]  >>{COLOR.YELLOW} TCP{COLOR.WHITE} | {COLOR.GREEN}{port} {COLOR.WHITE}/ {COLOR.YELLOW} {name} | {product} | {version} | {extrainfo}"
+
+
+def close(port):
+    return (
+        f"\r{COLOR.WHITE}[{COLOR.BLUE}{DATE}{COLOR.WHITE}] {COLOR.WHITE}[{COLOR.RED}CLOSE{COLOR.WHITE}] >>{COLOR.RED} {port} ",
+    )
+
+
+def login_data(
+    user,
+    password,
+    hostname,
+    port,
+) -> str:
     return f"""
 
         {COLOR.RED} 
@@ -19,6 +38,22 @@ def ssh_login_data(user, password, hostname, port) -> str:
         {COLOR.WHITE}            port     :{COLOR.YELLOW} {port}
 
                       """
+
+
+def prompt(
+    param,
+    prompt,
+    password=None,
+):
+    return f"{COLOR.WHITE}[{COLOR.BLUE}{DATE}{COLOR.WHITE}] {COLOR.WHITE}[{COLOR.RED}{prompt}{COLOR.WHITE}]  {COLOR.YELLOW}{param}{COLOR.WHITE}:{COLOR.YELLOW}{password}"
+
+
+def sucsess(param):
+    return f"{COLOR.WHITE}[{COLOR.BLUE}{DATE}{COLOR.WHITE}]{COLOR.GREEN} {param}"
+
+
+def failed(param):
+    return f"{COLOR.WHITE}[{COLOR.BLUE}{DATE}{COLOR.WHITE}]{COLOR.RED} {param}"
 
 
 OPTIONS = f"""                                                
@@ -100,3 +135,106 @@ TOOL_INFORMATION = f"""
 
         {COLOR.RED}   
         """
+
+TOP_PORTS = [
+    21,
+    22,
+    23,
+    25,
+    53,
+    80,
+    110,
+    111,
+    135,
+    139,
+    143,
+    443,
+    445,
+    993,
+    995,
+    1723,
+    3306,
+    3389,
+    5900,
+    8080,
+    8443,
+    10000,
+    50000,
+    32768,
+    32769,
+    32770,
+    32771,
+    49152,
+    49153,
+    49154,
+    49155,
+    49156,
+    49157,
+    49158,
+    49159,
+    49160,
+    49161,
+    49162,
+    49163,
+    49164,
+    49165,
+    49166,
+    49167,
+    49168,
+    49169,
+    49170,
+    49171,
+    49172,
+    49173,
+    49174,
+    49175,
+    49176,
+    49177,
+    49178,
+    49179,
+    49180,
+    49181,
+    49182,
+    49183,
+    49184,
+    49185,
+    49186,
+    49187,
+    49188,
+    49189,
+    49190,
+    49191,
+    49192,
+    49193,
+    49194,
+    49195,
+    49196,
+    49197,
+    49198,
+    49199,
+    49200,
+    49201,
+    49202,
+    49203,
+    49204,
+    49205,
+    49206,
+    49207,
+    49208,
+    49209,
+    49210,
+    49211,
+    49212,
+    49213,
+    49214,
+    49215,
+    49216,
+    49217,
+    49218,
+    49219,
+    49220,
+    49221,
+    49222,
+    49223,
+    49224,
+]
