+++ yaseer_tool/functions/ftp_login.py
@@ -1,11 +1,7 @@
-import datetime
 import os
 from ftplib import FTP
 
-import colorama
-
-color = colorama.Fore
-date = datetime.datetime.now().time()
+from consts import TOOL_INFORMATION, login_data, prompt
 
 
 def ftp_login(ftp_server, port, username, password, status):
@@ -15,7 +11,7 @@ def ftp_login(ftp_server, port, username, password, status):
     ftp = FTP()
     if status == "1,0":
         for user in open(username, "r").read().splitlines():
-            date = datetime.datetime.now().time()
+
             try:
                 ftp.connect(ftp_server, port)
                 ftp.login(user, password)
@@ -24,34 +20,16 @@ def ftp_login(ftp_server, port, username, password, status):
                     os.system("clear")
 
                 elif os.name == "nt":
-                    os.system("cls")
+                    os.system("cls")  #! unreachable
 
-                print(
-                    f"""
-{color.YELLOW}
-                         ░▒▓████████▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-{color.BLUE} 0Y Tool{color.YELLOW}                 ░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░░▒▓██████▓▒░░▒▓█▓▒░ 
-{color.GREEN} By : 0xYaser {color.YELLOW}           ░▒▓█▓▒░░▒▓█▓▒░  ░▒▓█▓▒░   ░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░  ░▒▓█▓▒░           
-{color.WHITE} version [0.1] {color.YELLOW}          ░▒▓████████▓▒░  ░▒▓█▓▒░   ░▒▓█▓▒░ 
-
-
-{color.WHITE}            username :{color.GREEN} {user} 
-{color.WHITE}            password :{color.GREEN} {password}
-{color.WHITE}            host     :{color.CYAN} {ftp_server}
-{color.WHITE}            port     :{color.YELLOW} {port}
-                      """
-                )
+                print(TOOL_INFORMATION)
+                print(login_data(user, password, ftp_server, port))
                 exit()
             except:
-                print(
-                    f"{color.WHITE}[{color.BLUE}{date}{color.WHITE}] {color.WHITE}[{color.RED}Bad Login{color.WHITE}]  {color.YELLOW}{user}{color.WHITE}:{color.YELLOW}{password}"
-                )
+                print(prompt(user, password, prompt="BAD LOGIN"))
     elif status == "0,1":
         for pas in open(password, "r").read().splitlines():
-            date = datetime.datetime.now().time()
+
             try:
                 ftp.connect(ftp_server, port)
                 ftp.login(username, pas)
@@ -61,34 +39,16 @@ def ftp_login(ftp_server, port, username, password, status):
                     os.system("clear")
 
                 elif os.name == "nt":
-                    os.system("cls")
+                    os.system("cls")  #! unreachable
 
-                print(
-                    f"""
-{color.YELLOW}
-                         ░▒▓████████▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-{color.BLUE} 0Y Tool{color.YELLOW}                 ░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░░▒▓██████▓▒░░▒▓█▓▒░ 
-{color.GREEN} By : 0xYaser {color.YELLOW}           ░▒▓█▓▒░░▒▓█▓▒░  ░▒▓█▓▒░   ░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░  ░▒▓█▓▒░           
-{color.WHITE} version [0.1] {color.YELLOW}          ░▒▓████████▓▒░  ░▒▓█▓▒░   ░▒▓█▓▒░ 
-
-                         
-{color.WHITE}            username :{color.GREEN} {username} 
-{color.WHITE}            password :{color.GREEN} {pas}
-{color.WHITE}            host     :{color.CYAN} {ftp_server}
-{color.WHITE}            port     :{color.YELLOW} {port}
-                      """
-                )
+                print(TOOL_INFORMATION)
+                print(login_data(user, password, ftp_server, port))
                 exit()
             except:
-                print(
-                    f"{color.WHITE}[{color.BLUE}{date}{color.WHITE}] {color.WHITE}[{color.RED}Bad Login{color.WHITE}]  {color.YELLOW}{username}{color.WHITE}:{color.YELLOW}{pas}"
-                )
+                print(prompt(user, password, prompt="BAD LOGIN"))
     elif status == "1,1":
         for user in open(username, "r").read().splitlines():
-            date = datetime.datetime.now().time()
+
             try:
                 password2 = open(password, "r").read().splitlines()[num]
                 ftp.connect(ftp_server, port)
@@ -99,29 +59,11 @@ def ftp_login(ftp_server, port, username, password, status):
                     os.system("clear")
 
                 elif os.name == "nt":
-                    os.system("cls")
+                    os.system("cls")  #! unreachable
 
-                print(
-                    f"""
-{color.YELLOW}
-                         ░▒▓████████▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-{color.BLUE} 0Y Tool{color.YELLOW}                 ░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░░▒▓█▓▒░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░░▒▓██████▓▒░░▒▓█▓▒░ 
-{color.GREEN} By : 0xYaser {color.YELLOW}           ░▒▓█▓▒░░▒▓█▓▒░  ░▒▓█▓▒░   ░▒▓█▓▒░ 
-                         ░▒▓█▓▒░░▒▓█▓▒░  ░▒▓█▓▒░           
-{color.WHITE} version [0.1] {color.YELLOW}          ░▒▓████████▓▒░  ░▒▓█▓▒░   ░▒▓█▓▒░ 
-
-                             
-{color.WHITE}            username :{color.GREEN} {user} 
-{color.WHITE}            password :{color.GREEN} {password2}
-{color.WHITE}            host     :{color.CYAN} {ftp_server}
-{color.WHITE}            port     :{color.YELLOW} {port}
-                      """
-                )
+                print(TOOL_INFORMATION)
+                print(login_data(user, password, ftp_server, port))
                 exit()
             except:
-                print(
-                    f"{color.WHITE}[{color.BLUE}{date}{color.WHITE}] {color.WHITE}[{color.RED}Bad Login{color.WHITE}]  {color.YELLOW}{user}{color.WHITE}:{color.YELLOW}{password2}"
-                )
+                print(prompt(user, password2, prompt="BAD LOGIN"))
             num = num + 1
