+++ yaseer_tool/functions/md5.py
@@ -1,15 +1,12 @@
-import datetime
 import hashlib
 import os
 
-
-from .. import COLOR, TOOL_INFORMATION
+from consts import COLOR, DATE, TOOL_INFORMATION
 
 
 def md5(md5_hasH, file):
 
     for i in open(file, "r").read().splitlines():
-        date = datetime.datetime.now().time()
 
         text = i
         md5_hash = hashlib.md5()
@@ -21,11 +18,11 @@ def md5(md5_hasH, file):
                 os.system("clear")
 
             elif os.name == "nt":
-                os.system("cls")
+                os.system("cls")  #! unreachable
             print(TOOL_INFORMATION)
-            print(f"{COLOR.WHITE}[{COLOR.BLUE}{date}{COLOR.WHITE}]{COLOR.GREEN} {i}")
+            print(f"{COLOR.WHITE}[{COLOR.BLUE}{DATE}{COLOR.WHITE}]{COLOR.GREEN} {i}")
             break
         else:
             print(
-                f"{COLOR.WHITE}[{COLOR.BLUE}{date}{COLOR.WHITE}]{COLOR.RED} {md5_hex}"
+                f"{COLOR.WHITE}[{COLOR.BLUE}{DATE}{COLOR.WHITE}]{COLOR.RED} {md5_hex}"
             )
